refactor(command): 重构命令处理器实现
- 移除独立的 CommandMatcher 结构体，将其功能整合到 CommandHandler 中
- 修复命令匹配逻辑，移除对 Matcher trait 的依赖
- 更新 HookHandler 实现，移除不必要的 Matcher trait
- 优化命令执行流程，改进参数解析和命令查找逻辑

+++ crates/puniyu_macros/src/adapter/hook.rs
@@ -11,7 +11,7 @@ struct HookArg {
 	name: Option<String>,
 	#[darling(rename = "type")]
 	r#type: Option<String>,
-	rank: Option<u64>,
+	rank: Option<u32>,
 }
 
 pub fn hook(args: TokenStream, item: TokenStream) -> TokenStream {
@@ -146,7 +146,7 @@ pub fn hook(args: TokenStream, item: TokenStream) -> TokenStream {
 				#hook_type
 			}
 
-			fn rank(&self) -> u64 {
+			fn rank(&self) -> u32 {
 				#hook_rank
 			}
